+++ cmf/cmf_core_src/upslope/connections/subsurfacefluxes.cpp
@@ -27,13 +27,13 @@ using namespace cmf::geometry;
 /************************************************************************/
 // Richards eq.
 const CellConnector connections::Richards_lateral::cell_connector=CellConnector(connections::Richards_lateral::connect_cells);
-void connections::Richards_lateral::connect_cells( Cell & cell1,Cell & cell2,int start_at_layer/*=0*/ )
+void connections::Richards_lateral::connect_cells( Cell & cell1,Cell & cell2,ptrdiff_t start_at_layer/*=0*/ )
 {
 	real w=cell1.get_topology().flowwidth(cell2);
 	if (w>0)
 	{
-		for (int i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
-			for (int j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
+		for (ptrdiff_t i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
+			for (ptrdiff_t j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
 				real ca=cell1.get_layer(i)->get_flow_crosssection(*cell2.get_layer(j));
 				if (ca>0)	{
 					real d=cell1.get_layer(i)->position.distanceTo(cell2.get_layer(j)->position);
@@ -133,13 +133,13 @@ real connections::Darcy::calc_q( cmf::math::Time t )
 
 }
 
-void connections::Darcy::connect_cells( Cell & cell1,Cell & cell2,int start_at_layer/*=0*/ )
+void connections::Darcy::connect_cells( Cell & cell1,Cell & cell2,ptrdiff_t start_at_layer/*=0*/ )
 {
 	real w=cell1.get_topology().flowwidth(cell2);
 	if (w>0)
 	{
-		for (int i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
-			for (int j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
+		for (ptrdiff_t i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
+			for (ptrdiff_t j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
 				real ca=cell1.get_layer(i)->get_flow_crosssection(*cell2.get_layer(j));
 				if (ca>0)	{
 					real d=cell1.get_layer(i)->position.distanceTo(cell2.get_layer(j)->position);
@@ -176,13 +176,13 @@ real connections::TopographicGradientDarcy::calc_q( cmf::math::Time t )
 
 }
 
-void connections::TopographicGradientDarcy::connect_cells( Cell & cell1,Cell & cell2,int start_at_layer/*=0*/ )
+void connections::TopographicGradientDarcy::connect_cells( Cell & cell1,Cell & cell2,ptrdiff_t start_at_layer/*=0*/ )
 {
 	real w=cell1.get_topology().flowwidth(cell2);
 	if (w>0)
 	{
-		for (int i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
-			for (int j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
+		for (ptrdiff_t i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
+			for (ptrdiff_t j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
 				real ca=cell1.get_layer(i)->get_flow_crosssection(*cell2.get_layer(j));
 				if (ca>0)	{
 					real d=cell1.get_layer(i)->position.distanceTo(cell2.get_layer(j)->position);
@@ -210,13 +210,13 @@ real connections::DarcyKinematic::calc_q( cmf::math::Time t )
 
 }
 
-void connections::DarcyKinematic::connect_cells( Cell & cell1,Cell & cell2,int start_at_layer/*=0*/ )
+void connections::DarcyKinematic::connect_cells( Cell & cell1,Cell & cell2,ptrdiff_t start_at_layer/*=0*/ )
 {
 	real w=cell1.get_topology().flowwidth(cell2);
 	if (w>0)
 	{
-		for (int i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
-			for (int j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
+		for (ptrdiff_t i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
+			for (ptrdiff_t j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
 				real ca=cell1.get_layer(i)->get_flow_crosssection(*cell2.get_layer(j));
 				if (ca>0)	{
 					real d=cell1.get_layer(i)->position.distanceTo(cell2.get_layer(j)->position);
